wip fix query regex verification
