+++ src/misc.ts
@@ -460,19 +460,3 @@ export const transformAddressToAddressInput = (data?: AddressType) => ({
   streetAddress1: data?.streetAddress1 || "",
   streetAddress2: data?.streetAddress2 || ""
 });
-
-export const flatten = (obj: unknown) => {
-  // Be cautious that repeated keys are overwritten
-
-  const result = {};
-
-  Object.keys(obj).forEach(key => {
-    if (typeof obj[key] === "object" && obj[key] !== null) {
-      Object.assign(result, flatten(obj[key]));
-    } else {
-      result[key] = obj[key];
-    }
-  });
-
-  return result;
-};
